Add new "diff ELF" mode for diffing stripped and non-stripped ELF
This is useful for diffing functions between two ELFs, in the situation
where the first one is stripped and the other one is non-stripped,
using start addresses to locate functions for the former and symbols
for the latter.

It's somewhat of a mix of the two existing modes, but I think making
a new mode is cleaner.

(My personal use case for this is diffing BotW using the original
executable and a ELF containing only the decompiled functions.
For various reasons, disassembling everything like the OoT decomp
project is annoying and inconvenient.)