Ignore immediate diffs for branches if they have identical targets
objdump prints branch targets using absolute addresses which means that
whenever the reference function and the decomp function do not have the
same base address, branches that are actually identical (with the
same PC-relative offset) will still be detected as being different.

This can be fixed by deriving the relative offset from the line number
and the absolute address that is given to us by objdump, and only
comparing the relative offsets instead.