Add an option to ignore address differences
This increases the signal-to-noise ratio of diffs in situations where
it is difficult to generate fully matching executables with all
functions and data being placed at the correct locations.

The new option only affects A64 diffs at the moment but extending it
to any other supported architecture shouldn't be too difficult.