Add a short alias for --source (-c) (#47)
After the addition of --source-old-binutils, passing --source now
requires writing the option in full. --so and other shorthands
do not work anymore because they are now ambiguous.

For the sake of convenience, this adds -c as an alias for --source
(c for sourCe or C/C++, I guess...)

Also adds a similar alias for --source-old-binutils (-C).