Fix the AArch64 register regex for FP regs followed by data width (#71)
* Fix the AArch64 register regex for FP regs followed by data width

FP registers may be followed by data width and number of elements,
e.g. V0.4S (instead of just V0). For simplicity, we don't actually
check if the combination of data width / number of elements makes sense
or list all the combinations in the regex -- instead, the regex simply
looks for one or two digits followed by a valid letter.

* Simplify the AArch64 register regex

https://github.com/simonlindholm/asm-differ/pull/71#pullrequestreview-853614276